Users cannot access locked projects, projects become locked if premium becomes free user and they have more than 1 project, users can choose one project to unlock if all projects are locked, all projects auto unlock if become premium user

+++ server/controllers/task/index.js
@@ -19,6 +19,10 @@ const getAllTasksForProject = async (req, res) => {
   if (!canAccess)
     return res.status(403).json({ error: 'You are not authorized' })
 
+  const isProjectLocked = await checkIfProjectLocked(id)
+  if (isProjectLocked)
+    return res.status(403).json({ error: 'Project has been locked' })
+
   try {
     const tasks = await Task.find({ project: id }).populate('project')
     return res.json(tasks)
@@ -33,6 +37,10 @@ const createTask = async (req, res) => {
     const canAccess = await checkIfOwnerOrEditor(req.email, project)
     if (!canAccess)
       return res.status(403).json({ error: 'You are not authorized' })
+
+    const isProjectLocked = await checkIfProjectLocked(project)
+    if (isProjectLocked)
+      return res.status(403).json({ error: 'Project has been locked' })
 
     const task = new Task({ item, nextAction, priority, project })
     await task.save()
@@ -51,6 +59,10 @@ const updateTask = async (req, res) => {
   const canAccess = await checkIfOwnerOrEditor(req.email, task.project)
   if (!canAccess)
     return res.status(403).json({ error: 'You are not authorized' })
+
+  const isProjectLocked = await checkIfProjectLocked(task.project)
+  if (isProjectLocked)
+    return res.status(403).json({ error: 'Project has been locked' })
 
   const { item, nextAction, priority } = req.body
   if (!item || !nextAction || !priority)
@@ -79,6 +91,10 @@ const deleteTask = async (req, res) => {
   if (!canAccess)
     return res.status(403).json({ error: 'You are not authorized' })
 
+  const isProjectLocked = await checkIfProjectLocked(task.project)
+  if (isProjectLocked)
+    return res.status(403).json({ error: 'Project has been locked' })
+
   task = await Task.deleteOne({ _id: id })
   res.send(task)
 }
@@ -102,6 +118,27 @@ const checkIfOwnerOrEditor = async (email, projectId) => {
   return email === project?.owner || project?.editors?.includes(email)
     ? true
     : false
+}
+const checkIfProjectLocked = async (projectId) => {
+  let project, user
+  try {
+    project = await Project.findById(projectId)
+  } catch {
+    return true
+  }
+  try {
+    user = await User.find({ email: project.owner })
+    user = user[0]
+  } catch {
+    return true
+  }
+  for (let p of user.ownProjects) {
+    // do .toString() incase it is passed as an objectId instead of a string
+    // remember: if we end up comparing object to object, obj1 === obj2 will never be the same even if
+    // they have the same contents. Hence, convert to string and do the comparision
+    if (p.project.toString() == projectId.toString()) return p.locked
+  }
+  return true
 }
 
 export {
